alot of stuff forgot to space it. added Guard Behaviour
list of improvements:
guards now patrol, investigate and chase the player based on suspision.
if the player touches a guard, they are instantly chased.
if a treasure is collected and a guard sees it gone, then it alerts all guards
if a guard goes into chase and is close to another guard, it alerts it as well.