added and modified code for split
the code that generates the treasure rooms and guards have been moved to their own scripts. the game manager calls the scripts in order to build them. this should hopefully begin the whole "customise" phase of the game but for now it creates the maze and treasure. next is to work on AI using API