Guards talking to each other via events
the treasure detection system worked like this: Guard spotting it, guard tells game manager, game manager tells all guards.

However, with the use of events, the guard reduces it's connection with the game manager by allowing it to send an event and alert the guards via listeners. if the game manager still needed to get that info, a listener can be added to the manager object and nothing would change.