modified Guard so there's no inheritance
the fact is that both standing and walking do the exact same things, just that standing only has one waypoint while walking has others. So it was kind of useless applying inheritance to the guard if they both do the same. this is to rectify it into something a little neater. the next steps into it is hopefully separate the guard code from the actual FSM, turning it into an independent package